requirements.txt
python-telegram-bot==20.7
transformers==4.37.2
torch==2.2.0
python-dotenv==1.0.1 